finalising merge - tests run but live test fails

+++ hcs-sxc-java-core/src/main/java/com/hedera/hcs/sxc/HCSCore.java
@@ -130,7 +130,7 @@ public enum HCSCore { // singleton implementation
     }
     
     
-    public HCSCore singletonInstanceDefault(int appId){
+    public HCSCore singletonInstanceDefault(long appId){
         if(this.applicationId!=-1) init(appId, "./config/config.yaml", "./config/.env");
         return INSTANCE;
     }
